pyproject.toml
[tool.poetry]
name = "accountbook"
version = "0.1.0"
description = ""
authors = ["Fisher Kuan <kuanfisher@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-auth-oauthlib = "^1.0.0"
google-cloud-bigquery = "^3.9.0"
gspread = "^5.8.0"
pandas = "^2.0.0"
pandas-gbq = "^0.19.1"
pre-commit = "^3.2.2"
pydantic = "^1.10.7"
pandera = "^0.14.5"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.black]
line-length = 120
